requirements.txt
# Core FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.21
alembic==1.13.1
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.6.4
pydantic-settings==2.1.0
email-validator==2.1.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# HTTP Requests
requests==2.31.0

# AI - Groq API (replacing Gemini)
groq==0.11.0
python-dotenv==1.0.1

# Optional: Keep Gemini as backup
google-generativeai==0.5.4
